finish

+++ src/services/OrderManagement/OrderManagement.Business/DTOs/OrderItem/OrderItemDto.cs
@@ -3,4 +3,4 @@ namespace OrderManagement.Business.DTOs.OrderItem;
 public record OrderItemDto(
     int ProductId,
     int Quantity,
-    int Price);
+    decimal Price);
